Move SelfType to its own file
