+++ pydantic/_internal/_model_construction.py
@@ -20,9 +20,10 @@ from . import _typing_extra
 from ._core_metadata import build_metadata_dict
 from ._core_utils import consolidate_refs
 from ._decorators import SerializationFunctions, ValidationFunctions
-from ._fields import Undefined, get_self_type, is_self_type
+from ._fields import Undefined
 from ._generate_schema import generate_config, model_fields_schema
 from ._generics import replace_types
+from ._self_type import get_self_type, is_self_type
 from ._utils import ClassAttribute, is_valid_identifier
 
 if typing.TYPE_CHECKING:
